feat: create copy text component
